requirements.txt

torch
torchvision
matplotlib
pillow
numpy
tqdm
